requirements.txt
Django==5.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1 